Create sidebar for desktop view
