+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import { ApolloWrapper } from "./ApolloWrapper";
-import BottomNav from "./components/landing/BottomNav";
+import Nav from "./components/landing/Nav";
 import Navbar from "./components/landing/Navbar";
 import "./globals.css";
 
@@ -20,11 +20,13 @@ export const metadata: Metadata = {
   description: "Lesson planning made simple",
 };
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const bottomNav = await Nav();
+
   return (
     <html lang="en">
       <body
@@ -33,7 +35,7 @@ export default function RootLayout({
         <ApolloWrapper>
           <Navbar />
           {children}
-          <BottomNav />
+          {bottomNav}
         </ApolloWrapper>
       </body>
     </html>
